Refactor to handle multiple/split packets
Previous iteration only handled one "frame" per tcp packet, but there
could obviously be more than one. The dissector would then only handle
the first frame, while ignoring everything that comes after. Or the tcp
packet could be cut in the middle of a very large silkroad frame,
causing the dissector to mess up the following packet as the header
would obviously be wrong (since it now contains data).

To handle the former, we can just run the dissector multiple times on
the same buffer, adding new tree information for each frame handled this
way. Each iteration, we keep track of how many bytes we've already used
(the offset) and work from there. In most cases, this will already get
us to the finish line, but this now makes us more open the second case.
To solve that, we can tell wireshark that it needs to desegment the
packets to create a full frame with exactly how many bytes we're
missing. It will then go ahead and append the next packet until we
finally finish.